requirements.txt
numpy
scipy
torch
torchvision
tqdm
Pillow
tensorflow
tensorboard
submitit
